updated ChangeLog; small fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@880 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CWorm_SendRecv.cpp
@@ -265,7 +265,7 @@ void CWorm::readPacket(CBytestream *bs, CWorm *worms)
 	// Check for a position hack (can occur from severe lag or screenshoting)
 	if (tGameInfo.iGameType == GME_HOST)  {
 
-		if(x > cServer->getMap()->GetWidth() || y > cServer->getMap()->GetHeight())
+		if(x > (short)cServer->getMap()->GetWidth() || y > (short)cServer->getMap()->GetHeight())
 			//vPos=cServer->FindSpot(); // Would not work because local worms are simulated on client's machine
 			cClient->SendDeath(iID, iID);
